% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_closest}
\alias{example_closest}
\title{closest intensities}
\format{
A matrix with 9 rows and 10 columns: each row represents an LED channel and each column represents an event
}
\usage{
example_closest
}
\description{
Matrix of closest intensities for example purposes. Generated from `target_irradiance`
}
\keyword{datasets}
