% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_intensities}
\alias{example_intensities}
\title{example intensities}
\format{
A matrix with 9 rows and 10 columns: each row represents an LED channel and each column represents an event
}
\usage{
example_intensities
}
\description{
Matrix of random intensities for example purposes
}
\keyword{datasets}
