% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_regime}
\alias{example_regime}
\title{regime matrix}
\format{
A matrix with 13 rows and 10 columns:
\describe{
 \item{time}{time in HH:MM:SS format}
 \item{hour}{The hour of the event}
 \item{minute}{The minute of the event}
 \item{second}{The second of the event}
 \item{380nm}{Intensity at 380nm LED channel}
 \item{400nm}{Intensity at 400nm LED channel}
 \item{etc.}{}
}
}
\usage{
example_regime
}
\description{
Example regime matrix
}
\keyword{datasets}
