% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helio.checkFormat.R
\name{helio.checkFormat}
\alias{helio.checkFormat}
\title{Check formatting of the heliospectra matrices}
\usage{
helio.checkFormat(check_matrix)
}
\arguments{
\item{check_matrix}{Matrix to be checked. Rows correspond to LEDs and columns are events / timepoints.}
}
\value{
Vector of booleans: TRUE = pass, FALSE = fail
}
\description{
Heliospectra DYNA matrices should have 9 rows (1 for each LED channel) and up to 150 columns (max number of events that Heliospectra DYNA can store)
}
\examples{
matrix_to_check <- LightFitR::target_irradiance
helio.checkFormat(target_irradiance)

}
