% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helio.checkWhite.R
\name{helio.checkWhite}
\alias{helio.checkWhite}
\title{Check that the 9th white channel is set to 0}
\usage{
helio.checkWhite(irradiance_matrix)
}
\arguments{
\item{irradiance_matrix}{Matrix of intended irradiances. rows = leds and columns = events}
}
\value{
Boolean: TRUE = pass, FALSE = fail
}
\description{
We currently cannot support programming the white (5700k) LED channel.
}
\keyword{internal}
