% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helio.csv_schedule.R
\name{helio.csv_schedule}
\alias{helio.csv_schedule}
\title{Format regime_matrix for csv output that Heliospectra lights can parse}
\usage{
helio.csv_schedule(regime_matrix, filename)
}
\arguments{
\item{regime_matrix}{Matrix containing light regime, as generated by makeRegime}

\item{filename}{Character. Filename to export to}
}
\value{
Matrix formatted for Heliospectra lights csv
}
\description{
Format regime_matrix for csv output that Heliospectra lights can parse
}
\examples{
tempfile_name = tempfile(fileext='.csv')
helio.csv_schedule(LightFitR::example_regime, tempfile_name)

}
