% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{helio.dyna.leds}
\alias{helio.dyna.leds}
\title{heliospectra DYNA LEDs}
\format{
A data frame with 9 rows and 3 columns:
\describe{
 \item{name}{Name of the LED channel}
 \item{wavelength}{Wavelength of the LED channel}
 \item{colour}{Colour of the LED channel}
}
}
\source{
<https://heliospectra.com/led-grow-lights/dyna/>
}
\usage{
helio.dyna.leds
}
\description{
Data about the heliospectra DYNA LED channels
}
\keyword{datasets}
