% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{helio.eventLimit}
\alias{helio.eventLimit}
\title{Maximum number of events}
\format{
Integer representing maximum allowable events
}
\usage{
helio.eventLimit
}
\description{
Maximum number of events programmable onto heliospectra
}
\keyword{datasets}
