% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.helioSchedule.R
\name{helio.writeSchedule}
\alias{helio.writeSchedule}
\title{Write the schedule to file that Heliospectra can parse}
\usage{
helio.writeSchedule(regime_matrix, filename, format = c("csv", "json"))
}
\arguments{
\item{regime_matrix}{Matrix containing light regime, as generated by makeRegime}

\item{filename}{Character. Filename to export to}

\item{format}{Character. Which format to export to? csv or json. Use extensions '.csv' or '.txt'}
}
\description{
Write the schedule to file that Heliospectra can parse
}
