% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.addWhiteZero.R
\name{internal.addWhiteZero}
\alias{internal.addWhiteZero}
\title{Add white (5700k) LED to intensities matricies after calculations}
\usage{
internal.addWhiteZero(intensities_matrix)
}
\arguments{
\item{intensities_matrix}{Matrix of intensities with 8 rows}
}
\value{
Matrix of intensities with 9 rows
}
\description{
Add white (5700k) LED to intensities matricies after calculations
}
