% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.calibCombine.R
\name{internal.calibCombine}
\alias{internal.calibCombine}
\title{Internal function to make a calibration dataframe from user-supplied vectors of calibration data}
\usage{
internal.calibCombine(
  calibration_leds,
  calibration_wavelengths,
  calibration_intensities,
  calibration_irradiances
)
}
\arguments{
\item{calibration_leds}{A numeric vector of LED values from calibration, mapping to intensities and irradiances (i.e. the same length)}

\item{calibration_wavelengths}{A numeric vector of wavelengths from calibration, corresponding to intensities and irradiances}

\item{calibration_intensities}{A numeric vector of intensities (heliospectra units) from calibration}

\item{calibration_irradiances}{A numeric vector of measured irradiances (any units, as long as it is consistently used) from calibration}
}
\value{
Correctly formatted dataframe of calibration data, for use in other functions
}
\description{
Internal function to make a calibration dataframe from user-supplied vectors of calibration data
}
\examples{
calib <- LightFitR::calibration
internal.calibCombine(calib$led, calib$wavelength, calib$intensity, calib$irradiance)

}
\keyword{internal}
