% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.checkNAs.R
\name{internal.checkNAs}
\alias{internal.checkNAs}
\title{Internal function to check matrix for NAs}
\usage{
internal.checkNAs(check_matrix)
}
\arguments{
\item{check_matrix}{Matrix to be checked. Rows correspond to LEDs and columns are events / timepoints.}
}
\value{
Boolean: TRUE = pass, FALSE = fail
}
\description{
Internal function to check matrix for NAs
}
\examples{
internal.checkNAs(LightFitR::example_intensities)

}
\keyword{internal}
