% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.makeTimes.R
\name{internal.makeTimes}
\alias{internal.makeTimes}
\title{Internal function. Takes vector of timepoints (POSICxt format) and converts into format readable by heliospectra.}
\usage{
internal.makeTimes(timeVector_POSICxt)
}
\arguments{
\item{timeVector_POSICxt}{Vector of schedule timepoints in POSICxt format}
}
\value{
Matrix of times. Each row corresponds to hours, minutes, seconds
}
\description{
Internal function. Takes vector of timepoints (POSICxt format) and converts into format readable by heliospectra.
}
\examples{
print(LightFitR::time_vector)
internal.makeTimes(LightFitR::time_vector)

}
\keyword{internal}
