% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.helio_json.R
\name{read.helio_json}
\alias{read.helio_json}
\title{Read a heliospectra script (json format) into a matrix.}
\usage{
read.helio_json(helio_script)
}
\arguments{
\item{helio_script}{File (.txt or .json) containing heliospectra regime script}
}
\value{
Matrix containing the regime encoded by the Heliospectra script
}
\description{
Read a heliospectra script (json format) into a matrix.
}
