% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_irradiance}
\alias{target_irradiance}
\title{target irradiances}
\format{
A matrix with 9 rows and 10 columns: each row represents an LED channel and each column represents an event
}
\usage{
target_irradiance
}
\description{
Matrix of random target irradiances for example purposes
}
\keyword{datasets}
