% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_vector}
\alias{time_vector}
\title{time vector}
\format{
A vector of length 10 with timepoints in POSIXct format
}
\usage{
time_vector
}
\description{
Example timepoints for events
}
\keyword{datasets}
