% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helio.disco.R
\name{helio.disco}
\alias{helio.disco}
\title{Write a disco schedule for fun!}
\usage{
helio.disco(filename, format = c("csv", "json"))
}
\arguments{
\item{filename}{Character. Filename to export to}

\item{format}{Character. Which format to export to? csv or json. Use extensions '.csv' or '.txt'}
}
\description{
Write a disco schedule for fun!
}
