% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helio.json_schedule.R
\name{helio.json_schedule}
\alias{helio.json_schedule}
\title{Format regime_matrix for json output that Heliospectra lights can parse}
\usage{
helio.json_schedule(regime_matrix, filename)
}
\value{
Character in json format that Heliospectra can parse
}
\description{
Format regime_matrix for json output that Heliospectra lights can parse
}
