% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.closestWavelength.R
\name{internal.closestWavelength}
\alias{internal.closestWavelength}
\title{Internal function. Find the closest wavelengths to the stated peaks}
\usage{
internal.closestWavelength(wavelength_vec, peak_wavelengths)
}
\arguments{
\item{wavelength_vec}{Vector of wavelengths that we have spectrometry data for}

\item{peak_wavelengths}{Vector of stated LED peaks}
}
\value{
Vector of wavelengths closest to the stated peaks
}
\description{
Internal function. Find the closest wavelengths to the stated peaks
}
\examples{
calib_wavelengths = LightFitR::calibration$wavelength
peaks = LightFitR::helio.dyna.leds$wavelength

internal.closestWavelength(calib_wavelengths, peaks)

}
