% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.tidyIntensities.R
\name{internal.tidyIntensities}
\alias{internal.tidyIntensities}
\title{Tidy intensities generated by algorithm so that lights can accept them}
\usage{
internal.tidyIntensities(intensities_matrix, calibration_intensities)
}
\arguments{
\item{intensities_matrix}{Matrix of predicted intensities}

\item{calibration_intensities}{A numeric vector of intensities (heliospectra units) from calibration}
}
\value{
Maxtrix of intensities which are integers and capped at the maximum possible intensity
}
\description{
Tidy intensities generated by algorithm so that lights can accept them
}
\keyword{internal}
