% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.tidyIntensities.R
\name{internal.tidyIntensities}
\alias{internal.tidyIntensities}
\title{Title}
\usage{
internal.tidyIntensities(intensities_matrix, calibration_intensities)
}
\arguments{
\item{intensities_matrix}{Matrix of predicted intensities}
}
\value{
Maxtrix of intensities which are integers and capped at the maximum possible intensity
}
\description{
Title
}
